--- packages/api-reference/src/helpers/deep-merge.ts
@@ -1,15 +1,0 @@
-/**
- * Deep merge for objects
- */
-export function deepMerge(source: Record<any, any>, target: Record<any, any>) {
-  for (const [key, val] of Object.entries(source)) {
-    if (val !== null && typeof val === 'object') {
-      target[key] ??= new val.__proto__.constructor()
-      deepMerge(val, target[key])
-    } else if (typeof val !== 'undefined') {
-      target[key] = val
-    }
-  }
-
-  return target
-}
