+++ packages/api-reference/src/hooks/useSidebar.ts
@@ -3,7 +3,7 @@ import { scrollToId } from '@/helpers/scroll-to-id'
 import { useNavState } from '@/hooks/useNavState'
 import { getHeadingsFromMarkdown, getLowestHeadingLevel } from '@/libs/markdown'
 import { getModels, hasModels, hasWebhooks } from '@/libs/openapi'
-import { isOperationDeprecated } from '@/libs/operation'
+import { isOperationDeprecated } from '@/libs/openapi'
 import { ssrState } from '@scalar/oas-utils/helpers'
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import type { Spec, Tag, TransformedOperation } from '@scalar/types/legacy'
